chore(table): move test and stories dirs
