+++ packages/table/stories/double-header-table.stories.tsx
@@ -9,7 +9,7 @@ import {
   Th,
   Thead,
   Tr,
-} from "../table"
+} from "../src"
 
 export default {
   title: "Table / Double Header",
