+++ packages/table/tests/table.test.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import { testA11y } from "@chakra-ui/test-utils"
-import { Table, TableCaption, Tbody, Td, Tfoot, Th, Thead, Tr } from "../table"
+import { Table, TableCaption, Tbody, Td, Tfoot, Th, Thead, Tr } from "../src"
 
 describe("<Table />", () => {
   it("should pass a11y test", async () => {
